+++ examples/templating_read/main_test.go
@@ -1,0 +1,12 @@
+// Copyright the templig contributors.
+// SPDX-License-Identifier: MPL-2.0
+
+package main
+
+import (
+	"testing"
+)
+
+func TestMainGood(t *testing.T) {
+	main()
+}