+++ funcs.go
@@ -5,6 +5,8 @@ package templig
 
 import (
 	"errors"
+	"io"
+	"os"
 	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
@@ -14,6 +16,7 @@ import (
 func templigFuncs() template.FuncMap {
 	result := sprig.TxtFuncMap()
 	result["required"] = required
+	result["read"] = readFile
 	return result
 }
 
@@ -24,4 +27,18 @@ func required(warn string, val any) (any, error) {
 	}
 
 	return val, nil
+}
+
+func readFile(fileName string) (any, error) {
+	file, err := os.Open(fileName)
+
+	if err != nil {
+		return "", err
+	}
+
+	defer func() { _ = file.Close() }()
+
+	content, readErr := io.ReadAll(file)
+
+	return string(content), readErr
 }